refactor: move DRM-related AMD code into its own file
Doing this as refactoring work before I start working on supporting
Intel GPUs.
